Move mock subpackage, use `go generate`
