+++ pkg/eks/api_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"github.com/aws/aws-sdk-go/service/eks/eksiface"
 	"github.com/aws/aws-sdk-go/service/sts/stsiface"
-	"github.com/weaveworks/eksctl/pkg/testutils/mocks"
+	"github.com/weaveworks/eksctl/pkg/eks/mocks"
 )
 
 type MockProvider struct {
