+++ pkg/eks/mocks/STSAPI.go
